BE-636 | Add pagination.proto, move to pkg/api
