+++ pkg/api/v1beta1/pools/http.go
@@ -1,4 +1,4 @@
-package types
+package pools
 
 import (
 	"strconv"
