--- proto/sqs/pools/pools.proto
@@ -1,17 +1,0 @@
-syntax = "proto3";
-package sqs.pools;
-
-import "gogoproto/gogo.proto";
-import "amino/amino.proto";
-import "google/protobuf/duration.proto";
-import "cosmos/base/v1beta1/coin.proto";
-import "cosmos/msg/v1/msg.proto";
-
-option go_package = "pools/types";
-
-// GetRequest is the request type for the Service.Get RPC method.
-message GetPoolsRequest {
-  repeated uint64 pool_id = 1;
-  uint64 min_liquidity_cap = 2;
-  bool with_market_incentives = 3;
-}
